requirements.txt
streamlit
python-dotenv
PyMuPDF
google-generativeai
PyPDF2